requirements.txt
pytsk3
libewf-python
bcrypt
psutil
pdf2image
pytesseract
olefile
tika
flask
flask_jwt_extended
flask_cors
python-evtx
PyMuPDF
pandas
scipy
ratsnlp
transformers
pywin32
python-dotenv
konlpy 
scikit-learn